Move screen shot tests to Android module
